refactor API structure by moving Admin and Role controllers to V1 namespace, and updating routes accordingly
